refactor: extract logger utilities to shared module and apply to all main scripts
Changes:
1. Created utils/logger_utils.py module:
   - Extracted all logger functions from product_info_export_main.py
   - Added setup_application_logging() for unified setup
   - Made logger utilities reusable across all scripts

2. Updated all main scripts to use logger utilities:
   - product_info_export_main.py
   - csv_export_main.py
   - csv_import_main.py
   - products_import_main.py
   - categories_import_main.py

3. Simplified main scripts:
   - Removed duplicate logger setup code
   - Single line call to setup_application_logging()
   - Consistent logging behavior across all applications

4. Enhanced test coverage:
   - Added 4 new tests for setup_application_logging()
   - Tests cover environment variable scenarios
   - Total: 290 tests passing

Benefits:
- DRY principle: Eliminated code duplication across 5 scripts
- Maintainability: Single source of truth for logger configuration
- Consistency: All applications use the same logging setup
- Testability: Shared utilities are easier to test
- Extensibility: Easy to add new logger types in the future
