+++ csv_import_main.py
@@ -4,19 +4,23 @@ import sys
 import logging
 from injector import Injector
 from batch.csv_import_processor import DataCsvImportProcessor
+from utils.logger_utils import setup_application_logging
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 def main():
     logger.info("Starting CSV Import Application...")
-    
+
+    # Setup application logging (Injector and SQLAlchemy)
+    setup_application_logging()
+
     injector = Injector()
 
     csv_import_processor = injector.get(DataCsvImportProcessor)
-    
+
     success = csv_import_processor.run_csv_import_process()
-    
+
     if success:
         logger.info("\nCSV import completed successfully!")
         return 0
